fix(mobile): native workspace header — real back button, title+branch subtitle, compose & menu toolbar
- flatten navigation: the (tabs) screen now lives directly in the root
  authenticated stack (intermediate workspace Stack removed), so pushing
  from home gets the native back chevron for free
- header config moves into the (tabs) layout via the Stack.Screen
  composition API (Stack.Title asChild + Stack.Toolbar must be DIRECT
  children — wrapper components are dropped by the matcher)
- title view is name + branch subtitle, width-capped so it never slides
  under the trailing buttons
- right side: native toolbar with a compose button (creates an ACP
  session, lands in the thread) and an ellipsis menu (copy branch,
  rename), sharing one glass capsule on iOS 26
