--- apps/mobile/screens/(authenticated)/workspace/[id]/components/WorkspaceToolbarMenu/index.ts
@@ -1,1 +1,0 @@
-export { WorkspaceToolbarMenu } from "./WorkspaceToolbarMenu";
